Some work on alternative Curve25519 implementations that support signing and verification.
 * Original code, doesn't have signing function
 * New JCE provider. Have private / public key generation, but no signature generation yet
 * Whispers signal library. Can do signatures and verification. Excluding JNA for the pure Java version

Sadly the signal library is not JPMS compliant, but I hope to finish the JCE provider when I can work out how the hell to do signature verification.

All of this will allow the client to finally make authenticated calls to the server, and potentially vice-versa (for live updates), and using the same private key used for the wireguard channel.